requirements.txt
# AtulayaAkriti Texture Rendering Tool Requirements
# Core Dependencies
streamlit==1.28.0
numpy>=1.24.0
opencv-python-headless==4.8.0.76
Pillow>=10.0.0
requests>=2.31.0

# AI/ML Dependencies
torch>=2.0.0
torchvision>=0.15.0

# Streamlit Components
streamlit-drawable-canvas>=0.9.3

# SAM (Segment Anything Model)
git+https://github.com/facebookresearch/segment-anything.git

# Additional utilities
scipy>=1.11.0
matplotlib>=3.7.0
scikit-image>=0.21.0
